+++ src/main.ts
@@ -1030,13 +1030,9 @@ if (!startup && !started) {
         }
     });
 
-    // Quit when all windows are closed, except on macOS. There, it's common
-    // for applications and their menu bar to stay active until the user quits
-    // explicitly with Cmd + Q.
+    // Quit when all windows are closed.
     app.on("window-all-closed", (): void => {
-        if (process.platform !== "darwin") {
-            app.quit();
-        }
+        app.quit();
     });
 
     if (!isSecondInstance) {